+++ Models/Transaction.cs
@@ -9,7 +9,7 @@ namespace WebBankingApp.Models
     public class Transaction
     {
         public Int32 Id { get; set; }
-        public Int32 Amount { get; set; }
+        public Decimal Amount { get; set; }
         public DateTime DateTime { get; set; }
         public virtual Share Share { get; set; }
     }
